# Improve header manipulation performance and usability