--- a/src/lib.rs
+++ b/src/lib.rs
@@ -1,3 +1,5 @@
+#![cfg_attr(feature = "unstable", feature(test))]
+
 #[macro_use]
 pub mod macros;