Example from the middle commit:

```Before:
test http::headers::bench::huge_to_utf8_bench     ... bench:   7,464,919 ns/iter (+/- 491,444)
test http::headers::bench::standard_to_utf8_bench ... bench:         695 ns/iter (+/- 82)

After:
test http::headers::bench::huge_to_utf8_bench     ... bench:   5,522,948 ns/iter (+/- 2,247,948)
test http::headers::bench::standard_to_utf8_bench ... bench:         445 ns/iter (+/- 26)
```

This is a marginal improvement to the "huge" benchmark, but a more typical case (standard_to_utf8) shows a nice ~25% improvement.

Also, to avoid retrieval of headers breaking due to case-sensitivity, all headers are normalized to lowercase as they are processed. The original case and ordering is maintained however and the HTTP request sent to the proxy target has all headers in the same order as received by Octopus from the client.